Refactoring and updating payment
