+++ tests/test_discount.py
@@ -1,9 +1,9 @@
 import unittest
 
-from customer import Customer
-from discount import Discount
-from order import Order
-from product import Product
+from models.customer import Customer
+from models.discount import Discount
+from models.order import Order
+from models.product import Product
 
 
 class TestDiscount(unittest.TestCase):
