+++ tests/test_payment.py
@@ -1,9 +1,9 @@
 import unittest
 
-from customer import Customer
-from order import Order
-from payment import Payment
-from product import Product
+from models.customer import Customer
+from models.order import Order
+from models.payment import Payment
+from models.product import Product
 
 
 class TestPayment(unittest.TestCase):
@@ -17,12 +17,6 @@ class TestPayment(unittest.TestCase):
         self.payment = Payment()
 
     def test_process_payment_success(self):
-        payment_info = {'card_number': '4111111111111111', 'expiry_date': '12/25', 'cvv': '123'}
-        self.assertTrue(self.payment.process_a_payment(self.order, payment_info))
+        payment_info = {'card_number': '4111111111111111', 'expiry_date': '12/25', 'cvv': '123', 'payment_status': 'success'}
+        self.assertTrue(self.payment.process_payment(self.order, payment_info))
         self.assertEqual(self.order.status, "PAID")
-
-    # disable this test case for now because the process_a_payment method is hard coded to succeed
-    # def test_process_payment_fail(self):
-    #     payment_info = {'card_number': '4111111111111111', 'expiry_date': '12/25', 'cvv': '123'}
-    #     self.assertFalse(self.payment.process_a_payment(self.order, payment_info))
-    #     self.assertEqual(self.order.status, "FAILED")
